Large scale redesign
We want to have a 4x4 grid of basepair images as well as the heat map and jmol
display. This is a lot to show on one page and this is my current attempt at it.
This is likely not the best solution, however, I will try it out for now. We add
such a 4x4 grid using my heatmap code as well. This moves the heatmap, 3D viewer
and images to the top and the more techincal information to the bottom.