Make the summary table interactive
We can now click on the summary table and see base combinations. It will display
all combinations with the case insensetive sequence. That is to say it will show
Aa and aA when clicking on AA. Right now this does not update the table to show
which are selected, and I don't know if we need to do that.