Refactor heat maps of different summary properties
This cleans up the code for it some what. However, there are still problems with
it, for example we should never show the count or resolution as -1 in a tool
tip. That's never good. In addition, we seem to have some problems with keeping
the count the same between the table and the heatmap.