Split up heatmap to several files
It is easier for me to deal with as several files. In addition, I now stop
fighting javascript to make callable objects. I have changed the API to produce
something more javascripty. I've also added the code from augment.js as a way to
do OO stuff. This simplifies some the OO stuff as I don't need to set prototypes
and constructors myself. This is an incomplete refactor as not everything works
yet but a better structure is here so I'll commit.