Allow cmd-click to deselect the last element
We can now use cmd-click to deselect the only selected element. However, this
change introduces a new behavior I'm not sure about. That now if we cmd-click a
range that overlaps what is already selected we do not append the whole
selection to the current selection, instead we select only the part of the range
that does not overlap the current selection.