Alter color scheme of exemplar table
We now use a light to dark blue color scheme. More blue is better. This is
consistent with the heatmap in that blue is better. However, I don't know what
cutoffs to use for bad so I don't have a set of colors that are red. Have a red
part of the scale would be good, but I'm not sure how to do it.