Add pdb file fragments to display
We use prealigned fragments for display now. This makes the superimposing
problem not an issue. This commit adds most, but not all of the fragments we
wish to display. Notably, we do not have fragments for things like: cWW and cwW,
only cWw. The filenames do not use case to determine if it is WW, wW or Ww.
Instead we use a flag of 0, 1 or 2 to show this.